TradeBot/requirements.txt
mysql-connector
aiogram==2.25.2
aiohttp==3.8.1
mysql-connector-python==8.0.29
requests==2.27.1